// src/generated/client/models/VacationRequest.ts
export interface VacationRequest {
    /**
     * ID of vacation request in UUID
     * @type {string}
     * @memberof VacationRequest
     */
    id?: string;
    /**
     * Person keycloak ID in UUID
     * @type {string}
     * @memberof VacationRequest
     */
    personId?: string;
    /**
     * Date of vacation start
     * @type {Date}
     * @memberof VacationRequest
     */
    startDate: Date;
    /**
     * Date of vacation end
     * @type {Date}
     * @memberof VacationRequest
     */
    endDate: Date;
    /**
     * Vacation days used for requested vacation
     * @type {number}
     * @memberof VacationRequest
     */
    days: number;
    /**
     * 
     * @type {VacationType}
     * @memberof VacationRequest
     */
    type: VacationType;
    /**
     * Message attached to request
     * @type {string}
     * @memberof VacationRequest
     */
    message: string;
    /**
     * ID of person who created status in UUID
     * @type {string}
     * @memberof VacationRequest
     */
    createdBy?: string;
    /**
     * Date time of vacation request registration
     * @type {Date}
     * @memberof VacationRequest
     */
    createdAt: Date;
    /**
     * Date time of vacation request update
     * @type {Date}
     * @memberof VacationRequest
     */
    updatedAt: Date;
}